+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## DEV
 
+ * Add the `removeField()` method to the palette manipulator (see contao/core-bundle#1668).
  * Allow to select a news reader in the news list (see contao/news-bundle#39).
  * Show the current page title and URL in the preview bar (see contao/core-bundle#1640).